+++ examples/basics/demo_custom_physics.py
@@ -28,6 +28,9 @@ the pseudo-inverse and proximal operators.
 """
 
 # %%
+
+# sphinx_gallery_tags = ["Basics"]
+
 from __future__ import annotations
 import deepinv as dinv
 import torch